Move serialization tests to airframeJVM
