+++ airframe/jvm/src/test/scala/wvlet/airframe/DesignSerializationTest.scala
@@ -1,0 +1,61 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package wvlet.airframe
+
+import java.io.{ByteArrayInputStream, ByteArrayOutputStream, ObjectInputStream, ObjectOutputStream}
+
+import DesignTest._
+
+object DesignSerializationTest {
+
+  implicit class RichDesign(d:Design) {
+    def serialize : Array[Byte] = {
+      val b = new ByteArrayOutputStream()
+      val oo = new ObjectOutputStream(b)
+      oo.writeObject(this)
+      oo.close()
+      b.toByteArray
+    }
+  }
+
+  def deserialize(b: Array[Byte]) : Design = {
+    val in = new ByteArrayInputStream(b)
+    val oi = new ObjectInputStream(in)
+    val obj = oi.readObject().asInstanceOf[Design]
+    obj.asInstanceOf[Design]
+  }
+}
+
+/**
+  *
+  */
+class DesignSerializationTest extends AirframeSpec {
+  import DesignSerializationTest._
+
+  "Design" should {
+    "be serializable" taggedAs("ser") in {
+      val b = d1.serialize
+      val d1s = deserialize(b)
+      d1s shouldBe (d1)
+    }
+
+    "serialize instance binding" taggedAs("ser1") in {
+      val d = Design.blanc.bind[Message].toInstance(Hello("world"))
+      val b = d.serialize
+      val ds = deserialize(b)
+      ds shouldBe (d)
+    }
+
+  }
+}
